requirements.txt
Mako==1.0.0
Markdown==2.4.1
MarkupSafe==0.23
Pillow==2.4.0
PyYAML==3.11
argparse==1.2.1
distribute==0.6.24
python-dateutil==2.2
six==1.7.2
wsgiref==0.1.2
